Made changes in data model class

+++ app/src/main/java/com/dev/pigeonproviderapp/datamodel/OrderDetailsResponseDatamodel.java
@@ -198,10 +198,10 @@ public class OrderDetailsResponseDatamodel {
     private String address;
     @SerializedName("long")
     @Expose
-    private Integer _long;
+    private Double _long;
     @SerializedName("lat")
     @Expose
-    private Integer lat;
+    private Double lat;
 
     public String getAddress() {
       return address;
@@ -211,19 +211,19 @@ public class OrderDetailsResponseDatamodel {
       this.address = address;
     }
 
-    public Integer getLong() {
+    public Double getLong() {
       return _long;
     }
 
-    public void setLong(Integer _long) {
+    public void setLong(Double _long) {
       this._long = _long;
     }
 
-    public Integer getLat() {
+    public Double getLat() {
       return lat;
     }
 
-    public void setLat(Integer lat) {
+    public void setLat(Double lat) {
       this.lat = lat;
     }
